requirements.txt
discord.py
psutil
graphviz
streamlit
discord-ext-voice-recv
faster-whisper
duckduckgo-search
PyNaCl
imageio-ffmpeg
sentence-transformers
scipy
numpy
flask-cors
python-dotenv
openai
chromadb
tiktoken
httpx
pillow
piper-tts
soundfile
pytest
pytest-asyncio
llama-cpp-python